add cosign verify (nominally)
it isn't pretty, and we've also avoided answering the question of how we
retrieve the checksum securely, or how we know we have a good version of
cosign, but it does a cosign verify at build time, and that's progress!

It's not for us to answer all the hard questions, just showing people
how to use these tools is taking a step in the right direction.

Signed-off-by: Kingdon Barrett <kingdon@weave.works>